[RENAME] : viewmodel 이름 변경 #151
