+++ presentation/src/main/java/com/indipage/ui/article/ArticleFragment.kt
@@ -27,10 +27,10 @@ import kotlin.math.abs
 class ArticleFragment : BindingFragment<FragmentArticleBinding>(R.layout.fragment_article) {
 
     private val viewModel by viewModels<ArticleViewModel>()
-    private val mainViewModel by activityViewModels<SignInViewModel>()
+    private val parentViewModel by activityViewModels<SignInViewModel>()
     override fun onStart() {
         super.onStart()
-        mainViewModel.getUser()
+        parentViewModel.getUser()
     }
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
