+++ presentation/src/main/java/com/indipage/ui/savedspace/SavedSpaceFragment.kt
@@ -25,10 +25,10 @@ class SavedSpaceFragment :
 
     private lateinit var adapter: SavedSpaceAdapter
     private val viewModel by viewModels<SavedSpaceViewModel>()
-    private val mainViewModel by activityViewModels<SignInViewModel>()
+    private val parentViewModel by activityViewModels<SignInViewModel>()
     override fun onStart() {
         super.onStart()
-        mainViewModel.getUser()
+        parentViewModel.getUser()
     }
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
