+++ presentation/src/main/java/com/indipage/ui/splash/SplashActivity.kt
@@ -1,14 +1,15 @@
-package com.indipage.ui
+package com.indipage.ui.splash
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.indipage.ui.signin.SignInActivity
-import com.indipage.ui.tutorial.TutorialActivity
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+
         startActivity(Intent(this, SignInActivity::class.java))
         finish()
     }
